refactor(chat): 고객센터 챗봇 로직 분리 및 공통 경로 상수 적용
